+++ src/presentation/controllers/signup.spec.ts
@@ -1,8 +1,6 @@
 import { SignUpController } from './singnup'
-import { MissingParamError } from '../errors/missing-param-error'
-import { InvalidParamError } from '../errors/invalid-param-error'
+import { MissingParamError, InvalidParamError, ServerError } from '../errors'
 import { EmailValidator } from '../protocols/email-validator'
-import { ServerError } from '../errors/server-error'
 
 interface SutTypes{
   sut: SignUpController
